refactor: split oversized repository, GraphView, AIHarness, and search files
Summary of changes:
- Refactored `src/db/repository.ts` into a modular `src/db/repository/` directory with sub-modules for entities, claims, notes, links, snapshots, and web cache.
- Defined `IRepository` interface and implemented dependency injection via `DbContext` and `useRepository` hook.
- Decomposed `src/features/graph/GraphView.tsx` by extracting layout algorithms, keyboard navigation, touch handling, and snapshot management into separate modules and hooks.
- Decomposed `src/features/ai/AIHarness.tsx` into `ChatView` and `SettingsWizard` components, and extracted logic into `useChat` and `useRateLimiter` hooks.
- Modularized `src/lib/search.ts` into `src/lib/search/` with separate files for Orama management, progressive search, FTS5 hydration, and external fetch handling.
- Addressed code review feedback by restoring accessibility regions, improving TypeScript visibility, and maintaining public API exports.
- Verified changes with unit tests and frontend verification scripts.
